backend/requirements.txt
blinker==1.8.2
click==8.1.7
contourpy==1.3.0
cycler==0.12.1
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
fonttools==4.53.1
greenlet==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.7
MarkupSafe==2.1.5
matplotlib==3.9.2
numpy==2.1.1
packaging==24.1
pandas==2.2.2
pillow==10.4.0
psycopg2-binary==2.9.9
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
six==1.16.0
SQLAlchemy==2.0.34
threadpoolctl==3.5.0
typing_extensions==4.12.2
tzdata==2024.1
Werkzeug==3.0.4
